feat(processor): Use JavaScriptSmCacheStacktraceProcessor by default for all projects
